requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.1.0
langchain>=0.3.25
langchain-core>=0.1.0
langchain-community>=0.3.26
langchain-huggingface>=0.0.3
langchain-openai>=0.3.24

# This is all you need for pinecone
langchain-pinecone>=0.2.8 

pypdf>=5.6.1
sentence-transformers>=4.1.0
langchain-huggingface>=0.0.3
huggingface-hub>=0.20.0
streamlit>=1.33.0
requests>=2.31.0